Cleanup new enums
